Develop

See merge request miotech-application/backend/mp2c-go!325

+++ internal/pkg/model/entity/point.go
@@ -3,7 +3,7 @@ package entity
 type Point struct {
 	Id      int    `json:"column:id"`
 	OpenId  string `gorm:"column:openid"`
-	Balance int    `json:"column:balance"`
+	Balance int64  `json:"column:balance"`
 }
 
 func (Point) TableName() string {
